# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "guru-core API",
    "description": "\nThe single endpoint the app talks to. A user states a goal; the service collects\nwhatever context exists, asks follow-up questions when something essential is missing,\nand produces three difficulty variants of a scheduled plan.\n\n**Authentication.** Every endpoint except `POST /v1/auth/google`, `GET /health` and the\npresigned `/v1/files/*` routes needs `Authorization: Bearer <jwt>`. Get the token from\n`POST /v1/auth/google`; it carries the user id and expires after `JWT_TTL_SECONDS`.\nThe team-facing `role-models` writes authenticate with `X-API-Key` instead.\n\n**Long-running work.** Plan generation, import parsing and calendar export run on a\nqueue, so the endpoints that start them return `202` with an id. Poll the resource\n(`GET /v1/plan-sessions/{id}`, `GET /v1/imports`) rather than expecting a synchronous\nresult — PostgreSQL is the source of truth for job state, Redis only caches it.\n\n**Errors.** Every failure returns the same envelope:\n`{\"error\": {\"code\": \"not_found\", \"message\": \"...\"}}`. The code is the snake_case name of\nthe domain error: `invalid_input` (422), `unauthorized` (401), `forbidden` (403),\n`not_found` (404), `conflict` (409), `reauth_required` (409), `rate_limited` (429).\n",
    "contact": {
      "name": "Quasar-Gang",
      "url": "https://github.com/Quasar-Gang/guru-core"
    },
    "license": {
      "name": "Proprietary — all rights reserved"
...
  "paths": {
...
    "/v1/imports/presign": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Step 1 of a file upload: reserve an import and get a direct-upload URL",
        "description": "First of the three calls that bring a file into the system.\n\nDeclare the file up front — name, MIME type and size — and this creates an import row\nin status `pending` and returns a presigned `upload_url`. Nothing is stored yet; the\nbytes never pass through this endpoint.\n\nThe full upload flow:\n\n1. `POST /v1/imports/presign` → `import_id` + `upload_url`.\n2. `PUT` the raw bytes to `upload_url` **as-is** (query string included), with the\n   same `Content-Type` you declared. Send no `Authorization` header: the URL carries\n   its own signature.\n3. `POST /v1/imports/{import_id}/complete` to queue the file for parsing.\n\nLimits and validation, all enforced here so a rejected file never gets uploaded:\n\n- **Maximum 20 MB** (`20971520` bytes). `size_bytes` is trusted at this step, so the\n  client must send the real size.\n- **`upload_url` is valid for 900 seconds (15 minutes)** — the returned `expires_in`.\n  Once it lapses the URL returns `403`; call presign again for a fresh one.\n- The format is decided now, from the file extension first and `content_type` only as\n  a fallback. Supported: `csv`, `xlsx`, `md`, `html`, `pdf`, `docx`, `ics`.\n- `filename` is reduced to a single path segment, so directories and `..` in the name\n  are stripped rather than rejected.\n\nAn import that is presigned but never completed simply stays `pending` and is never\nparsed; it is harmless to abandon.",
        "operationId": "presign_import_v1_imports_presign_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PresignRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "The new `import_id`, the URL to PUT the bytes to, the storage key, and the lifetime of that URL in seconds (900).",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PresignResult"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized` — missing, invalid or expired bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "`invalid_input` — `size_bytes` exceeds the 20 MB limit (20971520 bytes), `filename` is empty once stripped of path segments, or neither the extension nor `content_type` maps to a supported format.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` — too many requests for this user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "BearerJWT": []
          }
        ]
      }
    },
    "/v1/imports/{import_id}/complete": {
      "post": {
        "tags": [
          "imports"
        ],
        "summary": "Step 3 of a file upload: confirm the bytes landed and queue parsing",
        "description": "Call this once the `PUT` to the presigned URL has returned `200`.\n\nIt verifies the object really exists in storage, moves the import from `pending` to\n`queued`, and enqueues an `import.parse` job. **Parsing is asynchronous** — this\nreturns immediately, well before the file has been read.\n\nPoll `GET /v1/imports` until the row leaves `queued`:\n\n- `parsed` — success. `event_count` and `chunk_count` are now filled in and the\n  content is available to plan generation.\n- `failed` — the file could not be read; `error` carries the reason. The upload is\n  not retried automatically; presign and upload a corrected file.\n\nSkipping this call leaves a perfectly uploaded file stuck in `pending` forever.",
        "operationId": "complete_import_v1_imports__import_id__complete_post",
        "security": [
          {
            "BearerJWT": []
          }
        ],
        "parameters": [
          {
            "name": "import_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Import Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The import, now in status `queued`. Counts are still zero.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ImportView"
                }
              }
            }
          },
          "401": {
            "description": "`unauthorized` — missing, invalid or expired bearer token.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "`not_found` — no import with this id belongs to the caller. Imports are scoped per user, so another user's id looks the same as a missing one.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "`invalid_input` — nothing was found at the storage key for this import: the `PUT` to `upload_url` never happened, failed, or the URL had already expired. Presign again and re-upload.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "429": {
            "description": "`rate_limited` — too many requests for this user.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/files/{key}": {
      "put": {
        "tags": [
          "files"
        ],
        "summary": "Upload bytes to a presigned URL (step 2 of a file import)",
        "description": "The destination of the `upload_url` returned by `POST /v1/imports/presign`.\n\n**Do not construct this URL yourself.** Take `upload_url` verbatim, query string and\nall, and `PUT` the raw file bytes to it — no multipart wrapper, no JSON envelope.\n\nAuthorization comes entirely from the `exp`/`op`/`sig` query parameters, which are\nsigned for exactly this key, this operation and this expiry; send **no**\n`Authorization` header. The URL is good for 15 minutes and this route is exempt from\nthe per-minute rate limit, so a large upload cannot spend the user's request budget.\n\nSend the same `Content-Type` you declared at presign time; it is stored alongside the\nobject. A missing header falls back to `application/octet-stream`. The 20 MB limit was\nalready enforced at presign, so it is not re-checked here.\n\nA `200` only means the bytes landed. The import stays `pending` until you call\n`POST /v1/imports/{import_id}/complete`.\n\nThis route exists for the local storage backend; in a deployment backed by object\nstorage the presigned URL points at the bucket instead. Either way the client code is\nthe same: PUT to whatever `upload_url` says.",
        "operationId": "upload_file_v1_files__key__put",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Storage key, as embedded in the presigned URL.",
              "title": "Key"
            },
            "description": "Storage key, as embedded in the presigned URL."
          },
          {
            "name": "exp",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Unix timestamp after which the signature is refused.",
              "title": "Exp"
            },
            "description": "Unix timestamp after which the signature is refused."
          },
          {
            "name": "op",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Operation the signature was minted for; must be `put` here.",
              "title": "Op"
            },
            "description": "Operation the signature was minted for; must be `put` here."
          },
          {
            "name": "sig",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "HMAC signature over `op`, `key` and `exp`.",
              "title": "Sig"
            },
            "description": "HMAC signature over `op`, `key` and `exp`."
          }
        ],
        "responses": {
          "200": {
            "description": "Stored. The response body is empty.",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "403": {
            "description": "`forbidden` — the signature does not match, was minted for the other operation (a `get` URL cannot be used to `put`), or `exp` has passed. Presigned URLs live for 15 minutes; ask for a new one rather than editing the query string.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "files"
        ],
        "summary": "Download bytes from a presigned URL",
        "description": "Read back an object through a download URL the backend handed out.\n\nSame rules as the upload route: use the URL exactly as given, send no bearer token,\nand expect it to stop working 15 minutes after it was minted. The signature covers the\n`get` operation specifically, so an upload URL cannot be replayed as a download.\n\nThe body is always served as `application/octet-stream` regardless of the type the\nobject was stored with, so treat it as a binary download rather than something to\nrender inline. Like the upload route, it is exempt from the per-minute rate limit.",
        "operationId": "download_file_v1_files__key__get",
        "parameters": [
          {
            "name": "key",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Storage key, as embedded in the presigned URL.",
              "title": "Key"
            },
            "description": "Storage key, as embedded in the presigned URL."
          },
          {
            "name": "exp",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "description": "Unix timestamp after which the signature is refused.",
              "title": "Exp"
            },
            "description": "Unix timestamp after which the signature is refused."
          },
          {
            "name": "op",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "Operation the signature was minted for; must be `get` here.",
              "title": "Op"
            },
            "description": "Operation the signature was minted for; must be `get` here."
          },
          {
            "name": "sig",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "HMAC signature over `op`, `key` and `exp`.",
              "title": "Sig"
            },
            "description": "HMAC signature over `op`, `key` and `exp`."
          }
        ],
        "responses": {
          "200": {
            "description": "The object's bytes.",
            "content": {
              "application/json": {
                "schema": {}
              },
              "application/octet-stream": {}
            }
          },
          "403": {
            "description": "`forbidden` — the signature does not match, was minted for the other operation (a `get` URL cannot be used to `put`), or `exp` has passed. Presigned URLs live for 15 minutes; ask for a new one rather than editing the query string.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404": {
            "description": "`not_found` — the signature is valid but no object exists at that key: it was never uploaded, or has since been deleted.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "required": [
          "error"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "ImportView": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "source": {
            "type": "string",
            "title": "Source"
          },
          "format": {
            "type": "string",
            "title": "Format"
          },
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "event_count": {
            "type": "integer",
            "title": "Event Count",
            "default": 0
          },
          "chunk_count": {
            "type": "integer",
            "title": "Chunk Count",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "id",
          "source",
          "format",
          "filename",
          "status",
          "error",
          "created_at"
        ],
        "title": "ImportView",
        "description": "One import as seen by the API client."
      },
...
      "PresignRequest": {
        "properties": {
          "filename": {
            "type": "string",
            "title": "Filename"
          },
          "content_type": {
            "type": "string",
            "title": "Content Type"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes"
          }
        },
        "type": "object",
        "required": [
          "filename",
          "content_type",
          "size_bytes"
        ],
        "title": "PresignRequest"
      },
      "PresignResult": {
        "properties": {
          "import_id": {
            "type": "string",
            "format": "uuid",
            "title": "Import Id"
          },
          "upload_url": {
            "type": "string",
            "title": "Upload Url"
          },
          "storage_key": {
            "type": "string",
            "title": "Storage Key"
          },
          "expires_in": {
            "type": "integer",
            "title": "Expires In"
          }
        },
        "type": "object",
        "required": [
          "import_id",
          "upload_url",
          "storage_key",
          "expires_in"
        ],
        "title": "PresignResult"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "BearerJWT": {
        "type": "http",
        "description": "A JWT issued by `POST /v1/auth/google`. Send it as `Authorization: Bearer <token>`.",
        "scheme": "bearer"
      }
    }
  },
...
  "servers": [
...
}